Alter Display of User Groups
Make pages that display a list of users update the list as Ajax calls
are made to change the friend status of users. Lists of users include
'Friends', 'Friend Requests', and non-friends. The main index page
continues to display all users and so the status of users are simply changed.