Model Notifications on the Backend
Create the model/table to handle notifications to a user.
Add a column to the users table to track how many new notifications
a user has. Provide methods to manage tracking the new notifications.
Send notification when a friend request is made.
Test notifications on the backend TDD style.

Change the name of the *_create_friendships migrations (datetime on
previous filename was wrong).