Model Likes on the Backend
Create factories, tests, migration, and model for Like.
Add validations for new model as well as associations between
Like and User, Like and Post, and Like and Comment.
The Like associations with Post and Comment are polymorphic so
that both posts and comments can be liked.
Ensure that a User can't like the same Post or Comment multiple times.