Model Profile on the Frontend
Create controller/actions, views, and helper to enable a User to view
and edit their profile.
Include country_select gem to help in selecting a country for a profile.
Provide a user the ability to delete his/her Odinbook account, and
change the message for deleting a user account.
Write tests for editing and viewing a profile.