Commenting Improvements
Fix a problem which prevented the first comment of a post created with
an Ajax request from being displayed without a full page refresh.

Add a 'Comment' link to a post's display and use coffeescript to focus
on the comment input textfield when that link is clicked.
Also highlight a post when the comment input textfield is focused on.