Posts Authorization
Authorize users to be able to create post only on their newsfeed or
timeline or the timeline of their friends.
Authorize only the creator or receiver of a post to delete the post.
Hide the form for post creation on a user's timeline from other users
with whom they are not friends.

Use previously created ACCESS global constant in some old tests.
Use controller and request tests to authorize posts, TDD style.