Comments Authorization
Authorize users to comment on posts created or received by either them
or a friend of theirs.
Authorize only the creator of a comment (the commenter) to delete the
comment.
Use access_level? method to hide comment form and link from users
who are not friends with the current user.

Use controller and request tests to authorize comments, TDD style.