Friendships Authorization
Authorize only the receiver of a friend request (the friended of a
Friendship) to accept (ie, update) the Friendship.
Authorize only the two members (Users) of a Friendship to delete the
Friendship.

Refactor authorization from Posts, Comments, and Friendships
controllers into a new authorize_either_user! method.
Create factory for friendship and use it in tests to authorize
friendships, TDD style.