Sign Up Welcome
Move the sign up flash message and sending a sign up welcome email
to the application controller
Add a link to edit a new user's profile in the sign up message,
style it, and make the content of the sign up message HTML safe.
Add the sign up message and email for a user that signs up with
facebook (from omniauth callbacks controller) with the help of a
tested User class method.