Display Improvements
Make the unviewed notifications highlighted in the notifications index
display.
Add relevant message when a user has no friends, or friend requests, or
notifications to be displayed.
Make the Newsfeed and Timeline Post form small on page load and then
expand when it get focus.
Include various vendor prefixes for stylesheets.