Client Side Validation
Add client side validation on top of the previously implemented server
side validation.
Use the jQuery Validate plugin to validate the signup and edit profile
forms on the client side. For the signup validation, include a check
for unused email with an Ajax call to the server. For the edit profile
validation, use a custom validator to check the file size.
Use page specific javascript to run validation only where the forms
appear.
Modify view files and stylesheets to accomodate elements used to
display result of client side validation.