Infinite Scroll with Pagination
Set up the actions of the users controller for pagination and
modify and style views for pagination.
Use Ajax calls to new pages to implement infinite scroll in various
displays including: Newsfeed, Timeline, All Users, Friends, Find
Friends, and Friend Requests.
Hide Comment link in post view and display it with JavaScript.
Hide all but the latest 4 comments of a post with JavaScript and
provide a link to show hidden comments.
Move styling for shrinking and expanding post form to stylesheet
and add/remove class with JavaScript for the styling.