Realtime Updates
Use websocket-rails to add realtime update functionality.
Implement realtime updates for new notifications, and created or
destroyed likes, comments, and posts.
Generally, broadcast to a channel from the relevant controller,
listen from the client and send an Ajax request when a received
broadcast affects the client. Route the Ajax request back to the
controller and then a js.haml view updates the view as appropriate
to be in sync with the original user's action (creating or destroying
a resource.) Use controllers, routes, websocket channel, helpers,
and various views to accomplish the realtime updates.
Make the added or removed views for realtime updates fade in or out
slowly so users are aware of them.

Unfortunately the provided helpers for testing websocket-rails
Event Router and WebsocketRails Controllers could not be used
because only Broadcasting to a Channel (for like, comment and post
creation and deletion) and Sending an Event to a Specific Client or User (for new notification update) were used.