Static Pages
Create routes, controller, actions, layout and views for the following
static pages: About, Contact/Help, Privacy Policy, and Terms of Service.
Make content for Static pages (currently leaving out a few external
links). Adjust layouts to accomodate the static pages.

Create a footer with links to the static pages. Display the footer to a
user who is not logged in, or a logged in user who visits a static
page. Include links to the static pages in the sidebar.

Style the static pages and the footer that links to them.
is not logged in

Create Tests for static pages and use them, TDD style.