Improve Infinite Scroll with Pagination
Check not just for when the user scrolls towards the bottom of the
page, but also when the user expands the height of the viewport and
when deleted or removed content causes the end of currently viewed
page to come into view.
For changes in the height of the viewport, listen for changes to
the size of the window. For deleted content, trigger an event
(contentSizeReduced) when a Post, Comment, or Like on post is deleted
as well as when the Friendship status of a user is changed in a
non-index display. Check if more content should be loaded when
a scroll, resize, or contentSizeReduced event is detected.