Minor Appearance Changes
Expand the height of element containing user name in header and
_users partial, and reduce padding of comment input field to fully
display user name and comment entry.
Reduce the fade time when creating and destroying posts, comments,
and friendships. Use normal time (400ms) for creations and 300ms
for deletions.
Change color of Last Updated part of Privacy and Terms pages.